requirements.txt
streamlit>=1.30.0
httpx>=0.25.0
python-dotenv>=1.0.0
Pillow>=10.0.0
